ai-character-designer-agent/requirements.txt
# Google Agent Development Kit
google-adk
google-genai

# Required for Google Search tool
google-search-results

# Streamlit for web UI
streamlit
